pyproject.toml
[tool.poetry]
name = "mini-crm-repair"
version = "0.1.0"
description = "Mini-CRM Repair Requests (FastAPI, SQLAlchemy async, Streamlit)"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "app" }, { include = "ui" }]

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic = ">=2.0"
SQLAlchemy = "^2.0.36"
asyncpg = "^0.30.0"
alembic = "^1.13.2"
psycopg2-binary = "^2.9.9"
python-jose = {version = "^3.3.0"}
passlib = {version = "^1.7.4", extras=["bcrypt"]}
bcrypt = "==4.0.1"
python-multipart = "^0.0.17"
email-validator = "^2.2.0"
httpx = "^0.27.2"
streamlit = "^1.38.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
ruff = "^0.6.9"
black = "^24.8.0"
mypy = "^1.11.2"

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.ruff]
line-length = 100
target-version = "py313"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

